Add files via upload
This us the GRIDLAB-D input script. This is the file that will be run in the GRIDLAB-D Terminal. It will require player files as an input for each node in the system that can be generated by running the MATLAB script mat_input. This GLM file describes a 24.7 kV 60 charger system with each charger evenly spaced 360 feet apart.